Update chat state from voice input
